Updated EntityEnemy
Kevin please do not change the EntityEnemy's sprites to anything but the place-holders that they all ready were. Because we are going to have more then just that one speciffic enemy so we should have one default enemy that we simply edit to create a new one. Basically EntityEnemy is like a null enemy and not one that is going to be in the game, its like a base.